requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.10
langchain-core>=0.1.0
groq>=0.4.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
datasets>=2.14.0
huggingface-hub>=0.16.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0 